+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix issue where new pages did not have global script and library assignments
 * Remove script and library assignments on page deletion
+* Add support for renaming site page files
 
 ## 3.7.8